ehr-data-pipeline-backend/requirements.txt
unstructured[pdf]>=0.10.0
unstructured-client>=0.15.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
openai>=1.0.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
